Add files via upload
This dashboard provides an overview of the Indian restaurant industry in terms of ratings, votes, and cuisines. The dashboard shows the top 5 restaurants with the highest ratings and highest votes, as well as the restaurants with the most variety of cuisines and the highest ratings. The dashboard also shows the distribution of restaurant ratings by city and cuisine.
The purpose of this dashboard is to provide insights into the Indian restaurant industry that can be used to make strategic decisions about where to open new restaurants, what cuisines to offer, and how to market restaurants. The dashboard can also be used by consumers to find the best restaurants in their area.